pyproject.toml
[project]
name = "aforro-backend-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.2",
    "django-celery-results>=2.6.0",
    "djangorestframework>=3.16.1",
    "faker>=40.1.2",
    "hiredis>=3.3.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "redis>=7.1.0",
]
